requirements.txt
# E2E Object Detection Pipeline with Compression - Dependencies
# Python 3.10+ required (tested on Python 3.10.12)

# ============================================================================
# CORE DEPENDENCIES (Required)
# ============================================================================

# Deep Learning and Computer Vision
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Computer Vision and Image Processing
# Note: Compatible versions tested for stability
opencv-python==4.6.0.66
numpy==1.23.5
Pillow>=9.0.0

# Configuration Management
PyYAML>=6.0

# MLflow for Experiment Tracking
mlflow>=2.5.0

# ============================================================================
# SYSTEM DEPENDENCIES (Required for full functionality)
# ============================================================================

# Threading and Concurrency (Built-in Python modules used)
# - threading (built-in)
# - concurrent.futures (built-in)
# - queue (built-in)
# - asyncio (built-in)

# ============================================================================
# OPTIONAL DEPENDENCIES (Recommended)
# ============================================================================

# Data Processing and Analysis
pandas>=1.5.0
scipy>=1.9.0

# System Monitoring
psutil>=5.9.0

# Progress Bars and Logging
tqdm>=4.64.0

# GPU Monitoring (if using CUDA)
# nvidia-ml-py3>=7.352.0

# Advanced Image Processing
# scikit-image>=0.19.0
# matplotlib>=3.5.0

# ============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# ============================================================================

# Testing
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

# Code Formatting
# black>=22.0.0
# flake8>=5.0.0

# ============================================================================
# SYSTEM REQUIREMENTS
# ============================================================================

# For Ubuntu/Debian:
# sudo apt-get update
# sudo apt-get install -y python3-pip python3-dev
# sudo apt-get install -y libsm6 libxext6 libfontconfig1 libxrender1
# sudo apt-get install -y libglib2.0-0 libxrender1 libgomp1

# For CUDA support (optional, for GPU acceleration):
# Follow PyTorch installation guide: https://pytorch.org/get-started/locally/

# ============================================================================
# INSTALLATION NOTES
# ============================================================================

# 1. Install Python 3.10+
# 2. Create virtual environment: python3 -m venv venv
# 3. Activate: source venv/bin/activate (Linux/Mac) or venv\Scripts\activate (Windows)
# 4. Install: pip install -r requirements.txt
# 5. Run: python3 main.py --help

# Compatible versions tested on:
# - Ubuntu 22.04 LTS
# - Python 3.10.12
# - CPU and CUDA environments
